The RTL8139 driver is now fully WORKING!!!
I have made an (ugly) hack around the transmit problem (but then again, aren't
hacks ugly all the time?). I still don't know if this is the proper thing, but
the point is that it seems to work now. I don't know if it is my mistake or
the cards mistake that I cannot schedule more than one write at a time, but
I'm looking into this. I don't know the difference in performance between this
driver and the one in R5, but I'm guessing this one's still a bit slower. I'll
continue to hack the driver in order to maximize performance.


git-svn-id: http://svn.haiku-os.org/haiku/trunk/current@4364 a95241bf-73f2-0310-859d-f6bbb57e9c96