id by Chris Simmons
Just a small note about 'id'... All testing I've been able to perform on my machine in compiling it was done under BONE.

Seeing as how BeOS does not have true multi-user yet, it was impossible for me to simulate the usage of multiple users, groups, the group list functions.

It appears to work as intended, reporting the ID and NAME of the current and requested USERNAME, but further testing is required to be sure.

I've removed the -a argument, and implemented as best as I could this spec:

http://www.opengroup.org/onlinepubs/007904975/utilities/id.html

No effort has been made yet to deal with posix LOCALE issues, yet.


git-svn-id: http://svn.haiku-os.org/haiku/trunk/current@6364 a95241bf-73f2-0310-859d-f6bbb57e9c96