Added a server that caches stat data and directory contents. To use this feature run StatCacheServer somewhere in the background.
The first run of jam will be slower, but subsequent ones will usually
benefit. Only little, if the BeOS FS cache is able to hold the data,
but as soon as the number of files and directories jam has to inspect
hits the limit at which the cache drops data needed for the next run,
the performance difference will be very noticeable.
E.g. the time for a  "jam -n > /dev/null" in the root of a partially built
tree on my machine dropped from >1 min without the server to 13 secs
with it (not in the first run, of course).

Although tested quite a bit, I would still consider the feature
experimental.


git-svn-id: http://svn.haiku-os.org/haiku/trunk/current@9671 a95241bf-73f2-0310-859d-f6bbb57e9c96