Made sLocker static, as originally intended. If there is no be_app when the AppServerLink is created, it now falls into the debugger (maybe only for now?). Removed some debugging cruft.
BTW I changed the standard app_server resolution to 800x600 by
intention - at least for testing this is much more appropriate,
and can be reversed later again.
FWIW I think we should use 1024x768 as a standard or whatever
is the native display resolution. Use 640x480 as a fall back
only in safemode.


git-svn-id: http://svn.haiku-os.org/haiku/haiku/trunk@13130 a95241bf-73f2-0310-859d-f6bbb57e9c96