* We now use libdebug.so, which saves us a couple lines of code. * Print the debugger message in case the debugged team called debugger(),   and the description of an occurred exception respectively. * Properly keep gdb's internal thread list up to date. This fixes the   problem that an attached debugger always resumed all threads of the   debugged team. * A bit of cleanup of the attaching code.
Discovered an ugly bug: When invoking gdb with an executable -- regardless
of whether attaching gdb to a running process or just starting a debugger
session with the executable -- the executable is modified. I tested with
"crashing_app" and in the second debug session gdb fails to read the debug
info from the file and encounters a weird error when closing the file:
"No space left on device." I have the suspicion that this is another bug
in Axel's VM/file cache pet. :-P



git-svn-id: http://svn.haiku-os.org/haiku/haiku/trunk@13775 a95241bf-73f2-0310-859d-f6bbb57e9c96