* adds drawing commands from clients * adds concept of a current and a pending update session * marks dirty views being resized or moved
Some aspects of the design are buggy, others are slow, but I'm
approaching a good overview of what's needed and what problems
lurk in the details. In the end I hope to make things work fast
and correctly at all times. Adi or anybody else, feel free to
join the efforts.



git-svn-id: http://svn.haiku-os.org/haiku/haiku/trunk@15158 a95241bf-73f2-0310-859d-f6bbb57e9c96