* extracted the frame buffer memcpy routine from HWInterface.cpp   into a new frame_buffer_support.h * added blend32 routine for blending a certain color with   a scanline in the frame buffer * added "solid" versions of B_OP_ALPHA drawing with   B_ALPHA_OVERLAY alha function (blending on top of   a non-transparent background such as the frame buffer) * implemented an optimized shortcut for alpha blended   FillRect() in Painter * used the "packed" version of scanlines for shapes with   an outline thicker than 4 pixels (and filled shapes anyways),   this improves drawing speed when there are a few anti-aliased   pixels at the beginning of a scanline, then a solid fill and   some anti-aliased pixels at the end of the scanline. Such as   large letters.
To summarize: The alpha blending in Painter seems to be about
1.45 times faster than on BeOS R5 which benefits drawing large
shapes. For example, drawing a large alpha blended rounded rect
is 1.28 times faster on the Haiku app_server. On the other hand,
B_OP_COPY is quite tough to beat. It is currently 10 times faster
on R5. But a great deal seems to be caused by the Painter
rasterization algorithm itself, since commenting out the actual
drawing doesn't gain any speed.
The other useful experience I collected was that reading and
writing and over the PCI bus in the same loop really hurts
performance. It is actually faster (like 1.8 times!!) to allocate
a second buffer, read from frame buffer into that, doing the
blending at the same time, then writing the buffer back to the
screen.



git-svn-id: http://svn.haiku-os.org/haiku/haiku/trunk@15698 a95241bf-73f2-0310-859d-f6bbb57e9c96