When only the app version_info was contained in the resources/attribute we read from (and the system version_info missing) GetVersionInfo() always failed. Given that rc only writes the app version_info, mimeset never created a BEOS:APP_VERSION attribute (nor ones for supported types icons). The version attribute was usually created nevertheless, as a side effect of the subsequent setversion invocation.
Under Linux the attribute emulation can accidentially pick up the
attributes of an earlier deleted node that had the the same node ID as the
file in question, which in this case could cause an invalid
BEOS:APP_VERSION attribute (the app version_info part at least).

Now GetVersionInfo() doesn't fail anymore, when only one info could be
read (the other one is zeroed). Not sure, if that is what BeOS does, but
it shouldn't harm. This fixes bug #100.

Also made SetVersionInfo() zero out what couldn't be read before writing it
back.



git-svn-id: http://svn.haiku-os.org/haiku/haiku/trunk@16320 a95241bf-73f2-0310-859d-f6bbb57e9c96