* added a bunch of drivers by Siarzhuk Zharski -> a USB Mass Storgage module (SCSI bus manager add-on) -> a SiS 7018 AC97 driver (uses "old" audio driver interface) -> a USB Serial driver -> a USB Vision driver and media add-on (Haupauge WinTV USB) * moved R3MediaDefs.h from usb_audio driver to common place   headers/private/audio (also used by SiS 7018 driver) * added TV tuner frequency defines to data/settings/media/usb_vision
Some of these drivers are in unfinished state, particularily the
USB Vision driver, the VideoProducer doesn't seem to use it yet,
the USB Mass Storage module is known to work though, it also includes
add-ons for itself that add support for some "special" hardware, these
are not integrated with the Jamfile build system though.

Also I didn't much much time with the "CVS package" targets, the
ReadMes are not added, someone with more knowledge about this could
add them...

None of this stuff is added to the Haiku image, it is simply included
to be maintained in the Haiku tree from now on.

* fixed a bug in Video Producer sample inherited from Be Sample
  code - the timing from the time source was not really used,
  on some systems this could cause in the producer not waking
  up at the correct time if the system time and audio card time
  are drifting apart



git-svn-id: http://svn.haiku-os.org/haiku/haiku/trunk@17625 a95241bf-73f2-0310-859d-f6bbb57e9c96