BRoster::Launch() eventually launches the application in question in several steps: 1. early pre-registration with the registrar ("I wanna launch the    app, make sure noone interferes.") 2. load the app image 3. finish pre-registration with the registrar ("I have launched    the app, here is its team ID.") 4. start app main thread 5. send "on launch" messages to the app (argv, refs, others)
If the app is already running or being launched, 1. fails with a
conclusive error code and returns the team ID and the pre-registration
token of the app. Steps 2 - 4 are skipped and only the messages are
delivered using the team ID returned by 1.

This change fixes a race condition: The failed early pre-registration
request obviously cannot return the team ID, if the other thread
launching the app has not finished step 3 yet. Thus the argv/refs
message would not get delivered and Launch() would not return the
correct team ID.

Now we wait for the pre-registration to be finished in this case, using
the former _IsAppPreRegistered() mechanism, which already provided
such a waiting feature for one request. It has been extended to
accomodate an arbitrary number of waiting requests and renamed to
_IsAppRegistered().

This fixed bug #763.



git-svn-id: http://svn.haiku-os.org/haiku/haiku/trunk@18728 a95241bf-73f2-0310-859d-f6bbb57e9c96