This removes some of the accidental submissions from my last commit.
For those wondering "mark_as" is the directory for the MarkAsRead and MarkAsNew
Tracker add-ons I'm planning to add (for email.) If anyone is violently 
opposed to this idea, speak now.

The WorkspacesLayer change was mostly an update to existing and ifdef 0'd
code so it would compile under ifdef 1. I left those changes and just set the
ifdef back to 0 here.

The other changes below were debugging and build changes that definitely were
not needed in the repository.

The undocumented changes in the last submission include some new keyboard
shortcuts in ShowImage as well as various changes to the input server and
roster as part of my ongoing work on those. I will clean this up some more 
and make another submission in a few hours.

Note to self: never submit from a high-level directory...


git-svn-id: http://svn.haiku-os.org/haiku/haiku/trunk@19480 a95241bf-73f2-0310-859d-f6bbb57e9c96