Added support to ShowImage for sending an image to the Trash by scripting Tracker. For now it only works when the delete key is pressed. I suppose a menu item might be useful, but I can do that later if people really want it :)
When the image is trashed the next image in the directory is opened. If the
trashed image is the last in the directory the window is closed.

The code seems pretty solid as I even tested it with Tracker not running. It
also works fine in R5 (in fact that is where I tested it.)

As shown I would prefer to use the Tracker headers rather than redefining the
constant used for the trashing message. I'm already using Tracker's app MIME
type from the private headers.


git-svn-id: http://svn.haiku-os.org/haiku/haiku/trunk@19564 a95241bf-73f2-0310-859d-f6bbb57e9c96