New handy build system feature: It is now possible to pass a command line to jam that contains build targets and that will be executed by the build system after building the targets and replacing their occurrences in the command line by their paths. The keyword indicating such a command line is "run", targets are marked by a leading ":". E.g.:
  jam -q run ':<build>xres' -l :libtracker.so

This builds the xres tool for the host platform and libtracker.so for
Haiku, and afterwards lists the resources libtracker.so contains. Note,
that this feature requires using Haiku's jam version.

The functionality is implemented by the new RunCommandLine rule, which
can, of course, also be used in the UserBuildConfig.



git-svn-id: http://svn.haiku-os.org/haiku/haiku/trunk@20598 a95241bf-73f2-0310-859d-f6bbb57e9c96