Initial check-in of Joysticks preferences, courtesy of Oliver Ruiz Dorantes.
There is still work to be done.

I did some code formatting as well as created the Jamfile and rdef. I copied
the R5 Joysticks bitmap icons, but am working on a Haiku vector icon. For now I
just copied the Fonts icon for purposes of having something in the rdef.


git-svn-id: http://svn.haiku-os.org/haiku/haiku/trunk@21051 a95241bf-73f2-0310-859d-f6bbb57e9c96