Resolve a TODO and fix another ancient bug, #386. Print Screen is now handled by BWindow, no longer by the app_server. This should stop the "screen freeze" effect.
This adds a dependency on libpng.so and libz.so to libbe.so. The same
dependencies and the PNGDump code added here can be removed from the
app_server. I am just waiting for a code review of this before doing that.

This implementation still does not give the client a chance to handle it
differently.


git-svn-id: http://svn.haiku-os.org/haiku/haiku/trunk@25269 a95241bf-73f2-0310-859d-f6bbb57e9c96