* Add a DumpFontList test that dumps all families and styles with most BFont   infos (should all be defaults expect for a few).
This shows that with our default installation there are quite a few "condensed"
styles and also an "extra light" variation. The condensed ones should probably
get their own face flag. The "extra light" and other unhandled styles should
probably not be classified as a B_REGULAR_FACE as this could lead applications
to use them in a wrong way yielding unexpected results (i.e. firefox that
picks up the "extra light" style for it's interface font).

git-svn-id: http://svn.haiku-os.org/haiku/haiku/trunk@25630 a95241bf-73f2-0310-859d-f6bbb57e9c96