- Since r21336, BView::ScrollBy was checking the values against the ScrollBar ranges but ScrollBy is often called before updating the scroll range (ie: in ContainerWindow.cpp). IMO, the programatic ScrollBy method shouldn't depend on the ScrollBars ranges or state. The original fix in r21336 was apparently hiding other BScrollBar or BView bugs that have been fixed in the mean time.   The content was offseted when going back to list mode after moving icons on the left/up in icon mode. This fixes Tracker bug #2312.
- Revert and fix changes to ContainerWindow.cpp in r18481 (cvs 1.37). The condition was broken, but it wouldn't ScrollBy() 
anyway due to the previous problem. Fixing BView made the content autoscroll even if the lefttop corner of the extent was 
already visible.

- Probably unrelated, fix changes to ContainerWindow.cpp in r18993 (cvs 1.38). PoseView()->Bounds().left/top < 0 is expected, 
if for example in icon mode you move an icon close or crossing the left side of the window and then scroll left to adjust. 
This fix ResizeToFit that wouldn't scroll the view correctly in some cases.

So we had a Tracker Bug uncovering a BView fix that was hiding another Tracker bug, everything is fixed hopefully, phew :-)


git-svn-id: http://svn.haiku-os.org/haiku/haiku/trunk@26043 a95241bf-73f2-0310-859d-f6bbb57e9c96