Imported the nice disk usage tool "Guido" into the repo as "DiskUsage". Guido was originally written by Mike Steed. The code was freely available from his website at one point, which I am no longer able to find. I believe the license was a kind of public domain, at least compatible with our MIT/X11 license.
The original code was well written with it's own coding style which was not
very far from the Haiku coding style. I have hopefully adopted it completely
to our coding style. While I did this, I came across several places with
comments for a minor bug which I was able to resolve all. I also found ways
to simplify certain places and I elliminated all non-necessary "inter-class
knowledge" which was used to directly access members of other classes which
should have been private.

All in all, this should be a nice little addition, a useful tool which I found
handy a lot of times. If I should not have added this without first consulting
everyone, please voice your disagreement, and it can be removed it again. It
is not part of the image yet.


git-svn-id: http://svn.haiku-os.org/haiku/haiku/trunk@27357 a95241bf-73f2-0310-859d-f6bbb57e9c96