* When a device closes, delete the transfer semaphore, so that threads blocking   on it are unblocked and get an error. * Make fOpen volatile to prevent unwanted caching effects when checking it from   different threads. (?) * Check IsOpen() in the KeyboardDevice class in more acquire_sem_etc() return   cases, analogous to the MouseDevice class.
I am still getting a problem when relaunching input_server with the input_server
add-on thread that ioctl()s on a USB keyboard fd, which should have never fired
because it's a fake device from a KVM. After the first input_server instance is
gone, this thread keeps on busy looping in acquire_sem_etc()->switch_sem() from
within the ioctl() of the KeyboardDevice usb_hid driver. Still on it.


git-svn-id: http://svn.haiku-os.org/haiku/haiku/trunk@28350 a95241bf-73f2-0310-859d-f6bbb57e9c96