Works around problem with runtime loader described here:
http://www.freelists.org/post/haiku-development/Runtime-loader-problem

The problem seems to be related to loading/unloading add-ons linked against the
recent OpenSSL. Linking the E-mail preferences against the SSL libraries (when
USE_SSL is enable) fixes this as the SSL libraries are not unloaded then.

I am still investigating this but will have no more time this week so I am commiting this workaround until I look at it again.



git-svn-id: http://svn.haiku-os.org/haiku/haiku/trunk@29320 a95241bf-73f2-0310-859d-f6bbb57e9c96