* Rewrote BDragger's mouse tracking to be asynchronous using a message filter. Placed the filter in kits/shared since i plan to use it in other places like Tracker. Animated replicants like ActivityMonitor wont stall anymore when initiating the drag. (On the desktop, Tracker's mouse tracking still busy loops sometimes, that's next on my list).
I had asynchronous long click detection (one button mouse support) in it but decided to get rid of 
it, it adds unneeded complexity in the code and is even getting in your way sometimes (ex: now you can 
take your time to drag the dragger, you wont be interrupted). If we want to reimplement that 
someday it should be done system wide anyway (only Tracker and replicants have that 'feature' 
AFAIK).


git-svn-id: http://svn.haiku-os.org/haiku/haiku/trunk@29972 a95241bf-73f2-0310-859d-f6bbb57e9c96