Small tweaks regarding the scrollbars of terminal when used in fullscreen mode :
* if there is only one tab open, no functional change
* when there is two or more tabs open, display the scroll bar in the _ActiveTermView

It simply didn't make sense to not display the scrollbars in the _ActiveTermView since it was already shown in the other tabs. On the other hand, I can understand not showing the scrollbar when there is no tabview visible, as it mimic a real terminal.

That fixes bug reported in ticket #3300.


git-svn-id: http://svn.haiku-os.org/haiku/haiku/trunk@30597 a95241bf-73f2-0310-859d-f6bbb57e9c96