Implemented pretty much all of #4190. The Enter key was actually already set as a shortcut for fullscreen, but I guess the view was processing it first. So I just removed B_ENTER as an option for next image. Also Shift plus primary mouse was already working for moving the image.
Also I know it was discussed at length but I still wonder if all these options
for next and previous image are overkill. Now there is no way to scroll with
the keyboard. And there isn't an easy way to go to the next image with the
mouse (menus are a pain.) So I still think some more UI tweaking is needed to
avoid constant shifting between mouse and keyboard.

Finally I also added Command-1 as a shortcut for Original Size, but it does not
work. Anyone know why?


git-svn-id: http://svn.haiku-os.org/haiku/haiku/trunk@32343 a95241bf-73f2-0310-859d-f6bbb57e9c96