- Added back code for scrolling with the cursor keys, this time only when Shift is pressed.
- Took Humdinger's excellent advice from #4190 to open the Backgrounds
preferences from ShowImage instead of setting the background manually. This
required some tweaks to the Backgrounds preferences to accept RefsReceived in
the application and forward that to the view. In addition the placement is
set to scaled since otherwise it will keep the old placement. How this is
done is sort of hacky, but I'm going to call it an "internal protocol." This
also allowed me to remove the BackgroundImage code from ShowImage which was
just copied from Tracker (yuck.)


git-svn-id: http://svn.haiku-os.org/haiku/haiku/trunk@32424 a95241bf-73f2-0310-859d-f6bbb57e9c96