* Applied localization to the Installer.  * Some coding style fixes along the way.  * Added German translation.
NOTE: The Installer will have to offer a way to change the current locale! So
any translations which are done at this point will have to receive one more
update after I added this feature.

NOTE2: There is something wrong with generating catalogs and the update-all
jam target. If you don't do a clean build, the build won't run through, but it
is enough to build for example an image and then run update-all on your actual
partition profile.


git-svn-id: http://svn.haiku-os.org/haiku/haiku/trunk@35209 a95241bf-73f2-0310-859d-f6bbb57e9c96